pyproject.toml
[tool.poetry]
name = "jspr"
version = "0.1.1"
description = "A JSON Processor"
authors = ["vector-of-bool <vectorofbool@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
typing-extensions = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
PyYAML = "^5.4.1"
pytest-xdist = "^2.2.1"
mypy = "^0.902"
pylint = "^2.8.3"
yapf = "^0.31.0"
rope = "^0.19.0"
Sphinx = "^4.0.2"
pyright = "^0.0.13"
typing-extensions = "^4.0.1"

[tool.poetry.scripts]
jspr = "jspr.main:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
